backend/app/shared/requirements.txt
aws-lambda-powertools[all] ~= 2.43
aws-requests-auth ~= 0.4
aws-xray-sdk ~= 2.12
mypy-boto3-cloudwatch ~= 1.35
mypy-boto3-lambda ~= 1.35
mypy-boto3-logs ~= 1.35
mypy-boto3-ram ~= 1.35
mypy-boto3-secretsmanager ~= 1.35
mypy-boto3-ssm ~= 1.35
mypy-boto3-sts ~= 1.35
mypy_boto3_dynamodb ~= 1.35
mypy_boto3_stepfunctions ~= 1.35
nest-asyncio ~= 1.5
pydantic ~= 1.10
retry2 ~= 0.9
